Transfer to MS SQL database with logic fixes and CRUD operations on software controller and service added

+++ AutomatikaUsers.Model/UserSoftware.cs
@@ -7,9 +7,9 @@ namespace AutomatikaUsers.Model
     public class UserSoftware
     {
         public User User { get; set; }
-        public ulong UserId { get; set; }
+        public int UserId { get; set; }
 
         public Software Software { get; set; }
-        public ulong SoftwareId { get; set; }
+        public int SoftwareId { get; set; }
     }
 }
